+++ projects/plugins/jetpack/_inc/client/writing/index.jsx
@@ -45,7 +45,6 @@ export class Writing extends React.Component {
 			'carousel',
 			'copy-post',
 			'custom-css',
-			'google-fonts',
 			'latex',
 			'masterbar',
 			'markdown',
@@ -55,6 +54,7 @@ export class Writing extends React.Component {
 			'infinite-scroll',
 			'widgets',
 			'widget-visibility',
+			'blocks',
 		].some( this.props.isModuleFound );
 
 		if ( ! this.props.searchTerm && ! this.props.active ) {