+++ projects/plugins/jetpack/class.jetpack.php
@@ -2338,37 +2338,7 @@ class Jetpack {
 	 * @param array  $headers List of headers, in the format array( 'HeaderKey' => 'Header Name' ).
 	 */
 	public static function get_file_data( $file, $headers ) {
-		// Get just the filename from $file (i.e. exclude full path) so that a consistent hash is generated.
-		$file_name = basename( $file );
-
-		$cache_key = 'jetpack_file_data_' . JETPACK__VERSION;
-
-		$file_data_option = get_transient( $cache_key );
-
-		if ( ! is_array( $file_data_option ) ) {
-			delete_transient( $cache_key );
-			$file_data_option = false;
-		}
-
-		if ( false === $file_data_option ) {
-			$file_data_option = array();
-		}
-
-		$key           = md5( $file_name . serialize( $headers ) );
-		$refresh_cache = is_admin() && isset( $_GET['page'] ) && 'jetpack' === substr( $_GET['page'], 0, 7 );
-
-		// If we don't need to refresh the cache, and already have the value, short-circuit!
-		if ( ! $refresh_cache && isset( $file_data_option[ $key ] ) ) {
-			return $file_data_option[ $key ];
-		}
-
-		$data = get_file_data( $file, $headers );
-
-		$file_data_option[ $key ] = $data;
-
-		set_transient( $cache_key, $file_data_option, 29 * DAY_IN_SECONDS );
-
-		return $data;
+		return ( new Modules() )->get_file_data( $file, $headers );
 	}
 
 	/**