+++ projects/plugins/jetpack/extensions/shared/components/product-management-controls/toolbar-control.js
@@ -8,14 +8,15 @@ import formatCurrency from '@automattic/format-currency';
  */
 import { BlockControls } from '@wordpress/block-editor';
 import { MenuGroup, MenuItem, ToolbarDropdownMenu } from '@wordpress/components';
-import { useDispatch, useSelect } from '@wordpress/data';
-import { store as editPostStore } from '@wordpress/edit-post';
+import { useSelect } from '@wordpress/data';
 import { __, sprintf } from '@wordpress/i18n';
 import { check, update, warning } from '@wordpress/icons';
 
 /**
  * Internal dependencies
  */
+import useOpenBlockSidebar from './use-open-block-sidebar';
+import { getMessageByProductType } from './utils';
 import { store as membershipProductsStore } from '../../../store/membership-products';
 
 function getProductDescription( product ) {
@@ -64,31 +65,32 @@ function Product( { onClose, product, selectedProductId, setSelectedProductId } 
 	);
 }
 
-function NewProduct( { onClose } ) {
-	const isEditorSidebarOpened = useSelect( select =>
-		select( editPostStore ).isEditorSidebarOpened()
-	);
-	const { openGeneralSidebar } = useDispatch( editPostStore );
+function NewProduct( { onClose, productType } ) {
+	const openBlockSidebar = useOpenBlockSidebar();
 
 	const handleClick = event => {
 		event.preventDefault();
-		// Open the sidebar if not open
-		if ( ! isEditorSidebarOpened ) {
-			openGeneralSidebar( 'edit-post/block' );
-		}
-		const input = document.getElementById( 'new-product-title' );
-		if ( input !== null ) {
-			//Focus on the new product title input
-			input.focus();
-		}
+		openBlockSidebar();
+		setTimeout( () => {
+			const input = document.getElementById( 'new-product-title' );
+			if ( input !== null ) {
+				//Focus on the new product title input
+				input.focus();
+			}
+		}, 100 );
 		onClose();
 	};
 
-	return <MenuItem onClick={ handleClick }>{ __( 'Add a new subscription', 'jetpack' ) }</MenuItem>;
+	return (
+		<MenuItem onClick={ handleClick }>
+			{ getMessageByProductType( 'add a new product', productType ) }
+		</MenuItem>
+	);
 }
 
 export default function ProductManagementToolbarControl( {
 	products,
+	productType,
 	selectedProductId,
 	setSelectedProductId,
 } ) {
@@ -103,16 +105,16 @@ export default function ProductManagementToolbarControl( {
 		productDescription = getProductDescription( selectedProduct );
 	}
 	if ( selectedProductId && ! selectedProduct ) {
-		productDescription = __( 'Subscription not found', 'jetpack' );
+		productDescription = getMessageByProductType( 'product not found', productType );
 		subscriptionIcon = warning;
 	}
 
 	return (
-		<BlockControls group="block">
+		<BlockControls __experimentalShareWithChildBlocks group="block">
 			<ToolbarDropdownMenu
 				className="product-management-control-toolbar__dropdown-button"
 				icon={ subscriptionIcon }
-				label={ __( 'Select a plan', 'jetpack' ) }
+				label={ getMessageByProductType( 'select a product', productType ) }
 				text={ productDescription }
 			>
 				{ ( { onClose } ) => (
@@ -129,7 +131,7 @@ export default function ProductManagementToolbarControl( {
 							) ) }
 						</MenuGroup>
 						<MenuGroup>
-							<NewProduct onClose={ onClose } />
+							<NewProduct onClose={ onClose } productType={ productType } />
 						</MenuGroup>
 					</>
 				) }